chore(Heap): move sdk loader to different file
